+++ starter/src/main/java/de/fraunhofer/iosb/ilt/faaast/service/starter/App.java
@@ -21,8 +21,14 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.DeserializationFeature;
+import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
+import com.jayway.jsonpath.Configuration;
+import com.jayway.jsonpath.DocumentContext;
+import com.jayway.jsonpath.JsonPath;
+import com.jayway.jsonpath.spi.json.JacksonJsonNodeJsonProvider;
+import com.jayway.jsonpath.spi.mapper.JacksonMappingProvider;
 import de.fraunhofer.iosb.ilt.faaast.service.Service;
 import de.fraunhofer.iosb.ilt.faaast.service.config.ServiceConfig;
 import de.fraunhofer.iosb.ilt.faaast.service.dataformat.DeserializationException;
@@ -31,7 +37,8 @@ import de.fraunhofer.iosb.ilt.faaast.service.endpoint.http.HttpEndpointConfig;
 import de.fraunhofer.iosb.ilt.faaast.service.endpoint.opcua.OpcUaEndpointConfig;
 import de.fraunhofer.iosb.ilt.faaast.service.exception.InvalidConfigurationException;
 import de.fraunhofer.iosb.ilt.faaast.service.model.exception.ValidationException;
-import de.fraunhofer.iosb.ilt.faaast.service.model.validation.ValueTypeValidator;
+import de.fraunhofer.iosb.ilt.faaast.service.model.validation.ModelValidator;
+import de.fraunhofer.iosb.ilt.faaast.service.model.validation.ModelValidatorConfig;
 import de.fraunhofer.iosb.ilt.faaast.service.starter.cli.LogLevelTypeConverter;
 import de.fraunhofer.iosb.ilt.faaast.service.starter.logging.FaaastFilter;
 import de.fraunhofer.iosb.ilt.faaast.service.starter.util.ServiceConfigHelper;
@@ -39,25 +46,24 @@ import de.fraunhofer.iosb.ilt.faaast.service.util.ImplementationManager;
 import de.fraunhofer.iosb.ilt.faaast.service.util.LambdaExceptionHelper;
 import io.adminshell.aas.v3.model.AssetAdministrationShellEnvironment;
 import io.adminshell.aas.v3.model.impl.DefaultAssetAdministrationShellEnvironment;
-import io.adminshell.aas.v3.model.validator.ShaclValidator;
-import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
+import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-import org.apache.jena.shacl.ValidationReport;
-import org.apache.jena.shacl.lib.ShLib;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import picocli.CommandLine;
@@ -100,10 +106,17 @@ public class App implements Runnable {
     protected static final String ENV_MODEL_KEY = "model";
     protected static final String ENV_MODEL_FILE_PATH = envPath(ENV_FAAAST_KEY, ENV_MODEL_KEY);
     // environment
-    protected static final String ENV_PATH_SEPERATOR = ".";
+    protected static final String ENV_PATH_SEPERATOR = "_";
+    protected static final String JSON_PATH_SEPERATOR = ".";
     // model
     protected static final String MODEL_FILENAME_DEFAULT = "aasenvironment.*";
     protected static final String MODEL_FILENAME_PATTERN = "aasenvironment\\..*";
+    private static final Configuration JSON_PATH_CONFIG = Configuration
+            .builder()
+            .mappingProvider(new JacksonMappingProvider())
+            .jsonProvider(new JacksonJsonNodeJsonProvider())
+            .build()
+            .addOptions(com.jayway.jsonpath.Option.SUPPRESS_EXCEPTIONS);
 
     @Option(names = "--no-autoCompleteConfig", negatable = true, description = "Autocompletes the configuration with default values for required configuration sections. True by default")
     public boolean autoCompleteConfiguration = true;
@@ -128,8 +141,8 @@ public class App implements Runnable {
     @Option(names = "--emptyModel", description = "Starts the FA³ST service with an empty Asset Administration Shell Environment. False by default")
     public boolean useEmptyModel = false;
 
-    @Option(names = "--no-modelValidation", negatable = true, description = "Validates the AAS Environment. True by default")
-    public boolean validateModel = true;
+    @Option(names = "--no-validation", negatable = false, description = "Disables validation, overrides validation configuration in core configuration.")
+    public boolean noValidation = false;
 
     @Option(names = "--loglevel-faaast", description = "Sets the log level for FA³ST packages. This overrides the log level defined by other commands such as -q or -v.")
     public Level logLevelFaaast;
@@ -237,21 +250,38 @@ public class App implements Runnable {
     }
 
 
-    private void validateModelIfRequired(ServiceConfig config) {
-        if (validateModel) {
-            try {
-                AssetAdministrationShellEnvironment model = config.getPersistence().getInitialModel() == null
-                        ? EnvironmentSerializationManager.deserialize(config.getPersistence().getInitialModelFile()).getEnvironment()
-                        : config.getPersistence().getInitialModel();
-                validate(model);
-            }
-            catch (IOException e) {
-                throw new InitializationException("Unexpected exception while validating model", e);
-            }
-            catch (DeserializationException e) {
-                throw new InitializationException("Error loading model file", e);
-            }
+    private void validate(ServiceConfig config) {
+        if (noValidation) {
+            config.getCore().setValidationOnLoad(ModelValidatorConfig.NONE);
+            config.getCore().setValidationOnCreate(ModelValidatorConfig.NONE);
+            config.getCore().setValidationOnUpdate(ModelValidatorConfig.NONE);
         }
+        try {
+            AssetAdministrationShellEnvironment model = config.getPersistence().getInitialModel() == null
+                    ? EnvironmentSerializationManager.deserialize(config.getPersistence().getInitialModelFile()).getEnvironment()
+                    : config.getPersistence().getInitialModel();
+            if (!config.getCore().getValidationOnLoad().isEnabled()) {
+                LOGGER.info("ValidateOnLoad is disabled in core config, no validation will be performed.");
+                return;
+            }
+            LOGGER.debug("Validating model...");
+            LOGGER.debug("Constraint validation: {}", config.getCore().getValidationOnLoad().getValidateConstraints());
+            LOGGER.debug("ValueType validation: {}", config.getCore().getValidationOnLoad().getValueTypeValidation());
+            LOGGER.debug("IdShort uniqueness validation: {}", config.getCore().getValidationOnLoad().getIdShortUniqueness());
+            LOGGER.debug("Identifier uniqueness validation: {}", config.getCore().getValidationOnLoad().getIdentifierUniqueness());
+            ModelValidator.validate(model, config.getCore().getValidationOnLoad());
+            LOGGER.info("Model successfully validated");
+        }
+        catch (DeserializationException e) {
+            throw new InitializationException("Error loading model file", e);
+        }
+        catch (ValidationException e) {
+            throw new InitializationException(
+                    String.format("Model validation failed with the following error(s):%s%s",
+                            System.lineSeparator(),
+                            e.getMessage()));
+        }
+
     }
 
 
@@ -299,7 +329,6 @@ public class App implements Runnable {
             ServiceConfigHelper.autoComplete(config);
         }
         withModel(config);
-        validateModelIfRequired(config);
         try {
             ServiceConfigHelper.apply(config, endpoints.stream()
                     .map(LambdaExceptionHelper.rethrowFunction(
@@ -310,11 +339,12 @@ public class App implements Runnable {
             throw new InitializationException("Adding endpoints to config failed", e);
         }
         try {
-            config = ServiceConfigHelper.withProperties(config, getConfigOverrides());
+            config = ServiceConfigHelper.withProperties(config, getConfigOverrides(config));
         }
         catch (JsonProcessingException e) {
             throw new InitializationException("Overriding config properties failed", e);
         }
+        validate(config);
         if (!dryRun) {
             runService(config);
         }
@@ -442,7 +472,7 @@ public class App implements Runnable {
             LOGGER.info("Model: empty (default)");
         }
         LOGGER.info("Model validation is disabled when using empty model");
-        validateModel = false;
+        config.getCore().setValidationOnLoad(ModelValidatorConfig.NONE);
         config.getPersistence().setInitialModel(new DefaultAssetAdministrationShellEnvironment.Builder().build());
     }
 
@@ -494,31 +524,6 @@ public class App implements Runnable {
     }
 
 
-    private void validate(AssetAdministrationShellEnvironment aasEnv) throws IOException {
-        LOGGER.debug("Validating model...");
-        try {
-            ValueTypeValidator.validate(aasEnv);
-        }
-        catch (ValidationException e) {
-            throw new InitializationException(
-                    String.format("Model type validation failed with the following error(s):%s%s",
-                            System.lineSeparator(),
-                            e.getMessage()));
-        }
-        ShaclValidator shaclValidator = ShaclValidator.getInstance();
-        ValidationReport report = shaclValidator.validateGetReport(aasEnv);
-        if (!report.conforms()) {
-            ByteArrayOutputStream validationResultStream = new ByteArrayOutputStream();
-            ShLib.printReport(validationResultStream, report);
-            throw new InitializationException(
-                    String.format("Detailed model validation failed with the following error(s):%s%s",
-                            System.lineSeparator(),
-                            validationResultStream));
-        }
-        LOGGER.info("Model successfully validated");
-    }
-
-
     /**
      * Collects config overrides from environment and CLI parameters.
      *
@@ -555,6 +560,74 @@ public class App implements Runnable {
                             .collect(Collectors.joining(System.lineSeparator())));
         }
         return result;
+    }
+
+
+    /**
+     * Collects config overrides from environment and CLI parameters.
+     *
+     * @param config used to replace certain separators
+     *
+     * @return map of config overrides
+     */
+    protected Map<String, String> getConfigOverrides(ServiceConfig config) {
+        return replaceSeparators(config, getConfigOverrides());
+    }
+
+
+    private Map<String, String> replaceSeparators(ServiceConfig config, Map<String, String> configOverrides) {
+        Map<String, String> result = new HashMap<>();
+        DocumentContext document = JsonPath.using(JSON_PATH_CONFIG).parse(mapper.valueToTree(config));
+        configOverrides.forEach((k, v) -> {
+            List<String> pathParts = new LinkedList<>(Arrays.asList(k.split(ENV_PATH_SEPERATOR)));
+            String newPath = getNewPathRecursive(document, pathParts.get(0), pathParts, 1);
+            result.put(newPath, v);
+        });
+        return result;
+    }
+
+
+    private String getNewPathRecursive(DocumentContext document, String currPath, List<String> pathParts, int nextPartIndex) {
+        String jsonPath = String.format("$.%s", currPath);
+        JsonNode node = document.read(jsonPath);
+
+        if (nextPartIndex >= pathParts.size()) {
+            return node == null ? null : currPath;
+        }
+
+        String nextPart = pathParts.get(nextPartIndex);
+
+        if (node == null) {
+            return getNewPathRecursive(
+                    document,
+                    currPath + ENV_PATH_SEPERATOR + nextPart,
+                    pathParts,
+                    nextPartIndex + 1);
+        }
+        else {
+            String pathWithDot = getNewPathRecursive(
+                    document,
+                    currPath + JSON_PATH_SEPERATOR + nextPart,
+                    pathParts,
+                    nextPartIndex + 1);
+            String pathWithSeparator = getNewPathRecursive(
+                    document,
+                    currPath + ENV_PATH_SEPERATOR + nextPart,
+                    pathParts,
+                    nextPartIndex + 1);
+            if (Objects.nonNull(pathWithDot) && Objects.nonNull(pathWithSeparator)) {
+                throw new InitializationException(String.format(
+                        "Ambiguity between '%s' and '%s', please set properties through the CLI or a configuration file.",
+                        pathWithDot,
+                        pathWithSeparator));
+            }
+            if (Objects.isNull(pathWithDot) && Objects.isNull(pathWithSeparator)) {
+                throw new InitializationException(String.format(
+                        "Unresolvable environment variable found '%s'.",
+                        pathWithDot));
+            }
+            return Objects.nonNull(pathWithDot) ? pathWithDot : pathWithSeparator;
+        }
     }
 
     /**